standalone/requirements.txt
black==21.12b0
flake8==4.0.1
pre-commit==2.16.0
paho-mqtt==1.6.1
RestrictedPython==5.2